+++ service/controller/chart/resource/releasemaxhistory/delete.go
@@ -1,0 +1,8 @@
+package releasemaxhistory
+
+import "context"
+
+// EnsureDeleted is a no-op.
+func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}