+++ service/controller/chart/resource/releasemaxhistory/resource.go
@@ -1,0 +1,52 @@
+package releasemaxhistory
+
+import (
+	"github.com/giantswarm/helmclient/v4/pkg/helmclient"
+	"github.com/giantswarm/microerror"
+	"github.com/giantswarm/micrologger"
+	"k8s.io/client-go/kubernetes"
+)
+
+const (
+	Name = "releasemaxhistory"
+)
+
+type Config struct {
+	// Dependencies.
+	HelmClient helmclient.Interface
+	K8sClient  kubernetes.Interface
+	Logger     micrologger.Logger
+}
+
+type Resource struct {
+	// Dependencies.
+	helmClient helmclient.Interface
+	k8sClient  kubernetes.Interface
+	logger     micrologger.Logger
+}
+
+// New creates a new configured releasemaxhistory resource.
+func New(config Config) (*Resource, error) {
+	// Dependencies.
+	if config.HelmClient == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.HelmClient must not be empty", config)
+	}
+	if config.K8sClient == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.K8sClient must not be empty", config)
+	}
+	if config.Logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+	}
+
+	r := &Resource{
+		helmClient: config.HelmClient,
+		k8sClient:  config.K8sClient,
+		logger:     config.Logger,
+	}
+
+	return r, nil
+}
+
+func (r Resource) Name() string {
+	return Name
+}